requirements.txt
Flask==3.0.3
pywebview==5.1
pyinstaller==6.8.0
